Modify the system creator to include libpyramicio.so and gcc
The system creator script is modified so that it creates binaries for
the shared library libpyramicio.so; it also makes the system install
gcc, make and other similar tools on first boot. Thus, a developer
will be able to compile directly his applications on the board.

Sample applications are included in /opt/pyramic from now on.